requirements.txt
pytest==7.2.0
Appium-Python-Client==2.7.1